requirements.txt
Flask==3.0.3
speedtest-cli
pandas
requests
